requirements.txt
django-filter>=23.0
celery>=5.2.0
redis>=4.5.0
django-extensions>=3.2.0
drf-yasg>=1.21.0
django-phonenumber-field>=7.1.0
phonenumbers>=8.13.0
geopy>=2.3.0